requirement_work.txt
absl-py==0.12.0
astor==0.8.1
attrs==20.3.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
dill==0.3.3
future==0.18.2
gast==0.2.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
h5py==3.2.1
idna==2.10
importlib-metadata==4.0.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.16.6
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.12.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1